requirements.txt
numpy
pandas
tensorflow
scikit-learn
pillow
matplotlib
streamlit
transformers
torch
torchvision
torchaudio
timm